Separate tests into their own files
